added more people, some sounds, and some random props
There is a problem with the sound I put for the crowd being heard from everywhere, but that's probably fine.  Not sure if we want to keep the light music in, but I tried it out since Wendy suggested it.  I turned down the volume on most things that already had sound